Renamed to kekab-style all of the actions and reducers
